CORE-023: Move RAG helpers to runtime services
